AngularJS - Directives
AngularJS directives are used to extend HTML. These are special
attributes starting with ng- prefix. We're going to discuss following
directives −

ng-app − This directive starts an AngularJS Application.

ng-init − This directive initializes application data.

ng-model − This directive defines the model that is variable to be used
in AngularJS.

ng-repeat − This directive repeats html elements for each item in a
collection.