AngularJS - Custom Directives
Custom directives are used in AngularJS to extend the functionality of
HTML. Custom directives are defined using "directive" function. A custom
directive simply replaces the element for which it is activated.
AngularJS application during bootstrap finds the matching elements and
do one time activity using its compile() method of the custom directive
then process the element using link() method of the custom directive
based on the scope of the directive. AngularJS provides support to
create custom directives for following type of elements.

Element directives − Directive activates when a matching element is
encountered.

Attribute − Directive activates when a matching attribute is
encountered.

CSS − Directive activates when a matching css style is encountered.

Comment − Directive activates when a matching comment is encountered.